zalo zns

+++ app/Http/Resources/MarketingResource.php
@@ -24,7 +24,7 @@ class MarketingResource extends JsonResource
             'schedules'        => @(int)$this->schedules,
             'percent_schedule' => !empty($this->contact) ? round(($this->schedules / $this->contact) *100 , 2) : 0,
             'percent_order'    => !empty($this->contact) ? round(($this->orders / $this->contact) *100 , 2) : 0,
-            'avg'              => !empty($this->orders) ? round(($this->gross_revenue / $this->orders) , 2) : 0,
+            'avg'              => !empty($this->orders) ? (int) round(($this->gross_revenue / $this->orders), 2) : 0,
             'gross_revenue'    => @$this->gross_revenue,
         ];
         return $result;
